morebits: New Morebits.batchOperation class
unlink, batchdelete, batchprotect, etc. all use their own idiosyncratic
mechanisms for batch operation. Rather than unifying them by copy-pasting,
I thought it would be better to generalise the process into morebits.

The current procedures break the list of pages into chunks, and then every
1 second, they check whether a new batch of API requests should be started.
While this is infallible (it doesn't rely on the completion of the previous
chunk before the next one can start), it is crude, and it doesn't handle
failures well (if there is an API error, the percentage never reaches 100%,
possibly leaving the user confused).

This class counts the completion of each task by means of a callback.
Tolerance is allowed for a small number of workers that may not complete
for some reason (unhandled exception, for example). This means that an
"onComplete" callback/event cannot be provided, as it may not reliably
fire.

The old approach may have been a necessity brought about by the fragility
of the old screen-scraping code, where workers were more likely to raise an
error and fail without returning. However, its advantages may still be
relevant; if the pure counting-based approach used here is found to be
unreliable, it may be necessary to alter Morebits.batchOperation to use a
timing method instead.

The code is not perfect but it will be improved as needed.