batchundelete: Convert to use Morebits.batchOperation
Also including a few more "miscellaneous ameliorations" that I found in an
old local branch from many months ago.

We don't need the "MinCutOff" anymore; not really sure why it was ever a
pref. Morebits.batchOperation hard-codes this value to be 0.1 * the chunk
size, or 2, whichever is greater. Since the default chunk size is 50, the
MinCutOff remains 5, so most users won't see the difference.