batchdelete: Merge in some logic from delimages
Merging in the logic for displaying initial contributors, and unlinking
image instances.

This paves the way for a potential deletion of the delimages module. The
logic allowing the deletion of all images displayed on a page is not
available in batchdelete, but it is of questionable usefulness. If anyone
says they want it, we can add it back later.